requirements.txt
Pillow>=8.3.1
Django>=3.2.5
django-debug-toolbar>=3.2.1
django-simple-captcha>=0.5.14
